Applying SwiftUI Tutorial to the project
